pyproject.toml
[tool.poetry]
name = "riollama"
version = "0.1.0"
description = ""
authors = ["Hildeberto <hildeberto@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
rio-ui = "^0.9.1"
python-dotenv = "^1.0.1"
ollama = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
